Add feature to allow users to input project name
This commit implements a new feature that enables users to input a
project name through a form. Upon submission, a new Project instance is
created with the provided name and added to the projects array. This
array will be used later to update the DOM with the list of projects.

The feature includes appropriate error handling to prevent the addition
of empty or duplicate project names.